Fix tax calculations for free trials, one time fees, and record recurring payments.
Previously, a free trial would report taxes on the full amount of the subscription,
not $0.00. Resulting in incorrect tax totals in Tax Cloud. This also fixed how
one time fees are processed as well.

Lookups will be made without one time fee aggregatable items. This is to guarantee
that we get a proper rate percentage. Then, after the transaction is processed, an
additional request is made to the Lookup API containg the adjustments for the one
time fees. This allows for the rate percentage to be recorded for any future
payments, but for TaxCloud to properly report the correct tax, $0 in the case of
free trials.

Additionally, we now report the tax calculations using the AddTransactions API when
a child transaction is processed. This isn't perfect, as it is possible that the tax
rate might change between years for a long running subscription. And unfortunately,
since we rely upon the gateway's recurring cycle, it isn't currently possible to
correct the tax amount if it changes. In the future, we should implement an
update subscription request for changes like that.